requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
PyYAML==6.0
ratelimit==2.2.1
requests==2.27.1
tqdm==4.63.0
ujson==5.1.0
urllib3==1.26.8
